Migrate to defining a separate route helper module manually
To allow for `routes()` to be a plain function, and keep the macro
walled off a bit more.

Also update find_route to handle routes being nested lists. This will
let the `routes()` function be able to generate route lists that can be
dynamically created (e.g. a wrapper function that prepends middleware to
all of the routes nested)